+++ Client.Test/InfluxDBClientTest.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using InfluxDB3.Client.Config;
+using InfluxDB3.Client.Test.Utils;
 
 // ReSharper disable ObjectCreationAsStatement
 // ReSharper disable AssignNullToNotNullAttribute
@@ -34,7 +36,7 @@ public class InfluxDBClientTest
             {"INFLUX_ORG", "my-org"},
             {"INFLUX_DATABASE", "my-database"},
         };
-        SetEnv(env);
+        TestUtils.SetEnv(env);
 
         using var client = new InfluxDBClient();
 
@@ -50,24 +52,9 @@ public class InfluxDBClientTest
         Assert.That(ae.Message, Is.EqualTo("The URL of the InfluxDB server has to be defined"));
     }
 
-    private static void SetEnv(IDictionary<String, String> dict)
-    {
-        foreach (var entry in dict)
-        {
-            Environment.SetEnvironmentVariable(entry.Key, entry.Value, EnvironmentVariableTarget.Process);
-        }
-    }
-
     [TearDown]
     public void Cleanup()
     {
-        var env = Environment.GetEnvironmentVariables(EnvironmentVariableTarget.Process);
-        foreach (var key in env.Keys)
-        {
-            if (((string)key).StartsWith("INFLUX_"))
-            {
-                Environment.SetEnvironmentVariable((string)key, null, EnvironmentVariableTarget.Process);
-            }
-        }
+        TestUtils.CleanupEnv();
     }
 }